[CLEANUP] Organize XLIFF rules into category namespaces
- Move ConvertXliff1To2Fractor into the Transform namespace
- Move the source- and target-language validation rules into the CodeQuality namespace
- Mirror the new category structure in the rule tests and fix the ConvertXliff1To2Fractor test namespace that was missing the Tests segment
- Register the XLIFF package in the documentation generation script and regenerate the rules overview
- Note in the conversion rule that XLIFF 2.0 is only supported from TYPO3 v14 onwards
