+++ packages/fractor-xliff/rules-tests/CodeQuality/EnsureXliffHasTargetLanguageFractor/Fixtures/de.add-target-lang-v1.xlf.fixture
