+++ packages/fractor-xliff/rules-tests/CodeQuality/EnsureXliffHasTargetLanguageFractor/Fixtures/de.add-trglang-v2.xlf.fixture
