+++ packages/fractor-xliff/rules-tests/CodeQuality/EnsureXliffHasTargetLanguageFractor/config/fractor.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 use a9f\Fractor\Configuration\FractorConfiguration;
 use a9f\Fractor\ValueObject\Indent;
+use a9f\FractorXliff\CodeQuality\EnsureXliffHasTargetLanguageFractor;
 use a9f\FractorXliff\Configuration\XliffProcessorOption;
-use a9f\FractorXliff\ConvertXliff1To2Fractor;
 
 return FractorConfiguration::configure()
     ->withOptions([
         XliffProcessorOption::INDENT_CHARACTER => Indent::STYLE_TAB,
         XliffProcessorOption::INDENT_SIZE => 1,
     ])
-    ->withRules([ConvertXliff1To2Fractor::class]);
+    ->withRules([EnsureXliffHasTargetLanguageFractor::class]);
