+++ packages/fractor-xliff/rules-tests/Transform/ConvertXliff1To2Fractor/Fixtures/basic.xlf.fixture
