+++ packages/fractor-xliff/rules-tests/Transform/ConvertXliff1To2Fractor/Fixtures/with-target-and-approved.xlf.fixture
