+++ packages/fractor-xliff/rules-tests/Transform/ConvertXliff1To2Fractor/config/fractor.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 use a9f\Fractor\Configuration\FractorConfiguration;
 use a9f\Fractor\ValueObject\Indent;
 use a9f\FractorXliff\Configuration\XliffProcessorOption;
-use a9f\FractorXliff\EnsureXliffHasSourceLanguageFractor;
+use a9f\FractorXliff\Transform\ConvertXliff1To2Fractor;
 
 return FractorConfiguration::configure()
     ->withOptions([
         XliffProcessorOption::INDENT_CHARACTER => Indent::STYLE_TAB,
         XliffProcessorOption::INDENT_SIZE => 1,
     ])
-    ->withRules([EnsureXliffHasSourceLanguageFractor::class]);
+    ->withRules([ConvertXliff1To2Fractor::class]);
