+++ packages/fractor-xliff/rules/CodeQuality/EnsureXliffHasSourceLanguageFractor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace a9f\FractorXliff;
+namespace a9f\FractorXliff\CodeQuality;
 
 use a9f\Fractor\Contract\NoChangelogRequired;
 use a9f\FractorXliff\Contract\XliffFractorRule;
@@ -12,38 +12,33 @@ use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \a9f\FractorXliff\Tests\EnsureXliffHasTargetLanguageFractor\EnsureXliffHasTargetLanguageFractorTest
+ * @see \a9f\FractorXliff\Tests\CodeQuality\EnsureXliffHasSourceLanguageFractor\EnsureXliffHasSourceLanguageFractorTest
  */
-final class EnsureXliffHasTargetLanguageFractor implements XliffFractorRule, NoChangelogRequired
+final class EnsureXliffHasSourceLanguageFractor implements XliffFractorRule, NoChangelogRequired
 {
+    private const DEFAULT_SOURCE_LANGUAGE = 'en';
+
     public function refactor(XliffDocument $xliffDocument): ?XliffDocument
     {
-        $language = $this->extractLanguageFromFilename($xliffDocument->file->getFileName());
-        if ($language === null) {
-            return null;
-        }
-
         if ($xliffDocument->version === XliffVersion::V2_0) {
-            return $this->addV2TargetLanguage($xliffDocument, $language);
+            return $this->ensureV2SourceLanguage($xliffDocument);
         }
 
-        return $this->addV1TargetLanguage($xliffDocument, $language);
+        return $this->ensureV1SourceLanguage($xliffDocument);
     }
 
     public function getRuleDefinition(): RuleDefinition
     {
         return new RuleDefinition(
-            'Add target-language attribute to localized XLIFF files where the filename starts with a 2-letter ISO language code',
+            'Ensure XLIFF files have the required source-language (v1.x) or srcLang (v2.0) attribute',
             [new CodeSample(
                 <<<'CODE_SAMPLE'
-<!-- de.locallang.xlf -->
 <?xml version="1.0" encoding="UTF-8"?>
 <xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
-    <file source-language="en" datatype="plaintext" original="messages">
+    <file datatype="plaintext" original="messages">
         <body>
             <trans-unit id="label.hello">
                 <source>Hello</source>
-                <target>Hallo</target>
             </trans-unit>
         </body>
     </file>
@@ -51,14 +46,12 @@ final class EnsureXliffHasTargetLanguageFractor implements XliffFractorRule, NoC
 CODE_SAMPLE
                 ,
                 <<<'CODE_SAMPLE'
-<!-- de.locallang.xlf -->
 <?xml version="1.0" encoding="UTF-8"?>
 <xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
-    <file source-language="en" target-language="de" datatype="plaintext" original="messages">
+    <file source-language="en" datatype="plaintext" original="messages">
         <body>
             <trans-unit id="label.hello">
                 <source>Hello</source>
-                <target>Hallo</target>
             </trans-unit>
         </body>
     </file>
@@ -68,7 +61,7 @@ CODE_SAMPLE
         );
     }
 
-    private function addV1TargetLanguage(XliffDocument $xliffDocument, string $language): ?XliffDocument
+    private function ensureV1SourceLanguage(XliffDocument $xliffDocument): ?XliffDocument
     {
         $changed = false;
         $rootElement = $xliffDocument->document->documentElement;
@@ -83,39 +76,28 @@ CODE_SAMPLE
             if (($child->localName ?? '') !== 'file') {
                 continue;
             }
-            if ($child->getAttribute('target-language') !== '') {
-                continue;
+            if ($child->getAttribute('source-language') === '') {
+                $child->setAttribute('source-language', self::DEFAULT_SOURCE_LANGUAGE);
+                $changed = true;
             }
-
-            $child->setAttribute('target-language', $language);
-            $changed = true;
         }
 
         return $changed ? $xliffDocument : null;
     }
 
-    private function addV2TargetLanguage(XliffDocument $xliffDocument, string $language): ?XliffDocument
+    private function ensureV2SourceLanguage(XliffDocument $xliffDocument): ?XliffDocument
     {
         $rootElement = $xliffDocument->document->documentElement;
         if (! $rootElement instanceof \DOMElement) {
             return null;
         }
 
-        if ($rootElement->getAttribute('trgLang') !== '') {
+        if ($rootElement->getAttribute('srcLang') !== '') {
             return null;
         }
 
-        $rootElement->setAttribute('trgLang', $language);
+        $rootElement->setAttribute('srcLang', self::DEFAULT_SOURCE_LANGUAGE);
 
         return $xliffDocument;
-    }
-
-    private function extractLanguageFromFilename(string $filename): ?string
-    {
-        if (\preg_match('/^([a-z]{2})\./i', $filename, $matches) !== 1) {
-            return null;
-        }
-
-        return \strtolower($matches[1]);
     }
 }
