+++ packages/fractor-xliff/rules/CodeQuality/EnsureXliffHasTargetLanguageFractor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace a9f\FractorXliff;
+namespace a9f\FractorXliff\CodeQuality;
 
 use a9f\Fractor\Contract\NoChangelogRequired;
 use a9f\FractorXliff\Contract\XliffFractorRule;
@@ -12,33 +12,38 @@ use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \a9f\FractorXliff\Tests\EnsureXliffHasSourceLanguageFractor\EnsureXliffHasSourceLanguageFractorTest
+ * @see \a9f\FractorXliff\Tests\CodeQuality\EnsureXliffHasTargetLanguageFractor\EnsureXliffHasTargetLanguageFractorTest
  */
-final class EnsureXliffHasSourceLanguageFractor implements XliffFractorRule, NoChangelogRequired
+final class EnsureXliffHasTargetLanguageFractor implements XliffFractorRule, NoChangelogRequired
 {
-    private const DEFAULT_SOURCE_LANGUAGE = 'en';
-
     public function refactor(XliffDocument $xliffDocument): ?XliffDocument
     {
-        if ($xliffDocument->version === XliffVersion::V2_0) {
-            return $this->ensureV2SourceLanguage($xliffDocument);
+        $language = $this->extractLanguageFromFilename($xliffDocument->file->getFileName());
+        if ($language === null) {
+            return null;
         }
 
-        return $this->ensureV1SourceLanguage($xliffDocument);
+        if ($xliffDocument->version === XliffVersion::V2_0) {
+            return $this->addV2TargetLanguage($xliffDocument, $language);
+        }
+
+        return $this->addV1TargetLanguage($xliffDocument, $language);
     }
 
     public function getRuleDefinition(): RuleDefinition
     {
         return new RuleDefinition(
-            'Ensure XLIFF files have the required source-language (v1.x) or srcLang (v2.0) attribute',
+            'Add target-language attribute to localized XLIFF files where the filename starts with a 2-letter ISO language code',
             [new CodeSample(
                 <<<'CODE_SAMPLE'
+<!-- de.locallang.xlf -->
 <?xml version="1.0" encoding="UTF-8"?>
 <xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
-    <file datatype="plaintext" original="messages">
+    <file source-language="en" datatype="plaintext" original="messages">
         <body>
             <trans-unit id="label.hello">
                 <source>Hello</source>
+                <target>Hallo</target>
             </trans-unit>
         </body>
     </file>
@@ -46,12 +51,14 @@ final class EnsureXliffHasSourceLanguageFractor implements XliffFractorRule, NoC
 CODE_SAMPLE
                 ,
                 <<<'CODE_SAMPLE'
+<!-- de.locallang.xlf -->
 <?xml version="1.0" encoding="UTF-8"?>
 <xliff version="1.2" xmlns="urn:oasis:names:tc:xliff:document:1.2">
-    <file source-language="en" datatype="plaintext" original="messages">
+    <file source-language="en" target-language="de" datatype="plaintext" original="messages">
         <body>
             <trans-unit id="label.hello">
                 <source>Hello</source>
+                <target>Hallo</target>
             </trans-unit>
         </body>
     </file>
@@ -61,7 +68,7 @@ CODE_SAMPLE
         );
     }
 
-    private function ensureV1SourceLanguage(XliffDocument $xliffDocument): ?XliffDocument
+    private function addV1TargetLanguage(XliffDocument $xliffDocument, string $language): ?XliffDocument
     {
         $changed = false;
         $rootElement = $xliffDocument->document->documentElement;
@@ -76,28 +83,39 @@ CODE_SAMPLE
             if (($child->localName ?? '') !== 'file') {
                 continue;
             }
-            if ($child->getAttribute('source-language') === '') {
-                $child->setAttribute('source-language', self::DEFAULT_SOURCE_LANGUAGE);
-                $changed = true;
+            if ($child->getAttribute('target-language') !== '') {
+                continue;
             }
+
+            $child->setAttribute('target-language', $language);
+            $changed = true;
         }
 
         return $changed ? $xliffDocument : null;
     }
 
-    private function ensureV2SourceLanguage(XliffDocument $xliffDocument): ?XliffDocument
+    private function addV2TargetLanguage(XliffDocument $xliffDocument, string $language): ?XliffDocument
     {
         $rootElement = $xliffDocument->document->documentElement;
         if (! $rootElement instanceof \DOMElement) {
             return null;
         }
 
-        if ($rootElement->getAttribute('srcLang') !== '') {
+        if ($rootElement->getAttribute('trgLang') !== '') {
             return null;
         }
 
-        $rootElement->setAttribute('srcLang', self::DEFAULT_SOURCE_LANGUAGE);
+        $rootElement->setAttribute('trgLang', $language);
 
         return $xliffDocument;
+    }
+
+    private function extractLanguageFromFilename(string $filename): ?string
+    {
+        if (\preg_match('/^([a-z]{2})\./i', $filename, $matches) !== 1) {
+            return null;
+        }
+
+        return \strtolower($matches[1]);
     }
 }
